+++ web-app/src/app/(landing)/(auth)/forgot-password/data.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { emailSchema } from "@/app/(auth)/data";
+import { emailSchema } from "@/lib/better-auth/data";
 import { FormData } from "@/lib/form";
 
 const forgotPasswordFormSchema = z.object({
