+++ web-app/src/app/(landing)/(auth)/signup/data.ts
@@ -5,7 +5,7 @@ import {
   emailSchema,
   nameSchema,
   passwordSchema,
-} from "@/app/(auth)/data";
+} from "@/lib/better-auth/data";
 import { FormData } from "@/lib/form";
 
 const signUpFormSchema = z
